webapp/requirements.txt
fastapi>=0.104
uvicorn>=0.24
jinja2>=3.1
python-multipart>=0.0.6
pyjwt>=2.8
passlib[bcrypt]>=1.7
cryptography>=41.0
pydantic>=2.0
httpx>=0.25
stripe>=7.0.0
python-dotenv>=1.0.0
slowapi>=0.1.9
sqlcipher3-binary>=0.5.2
psutil>=5.9
